Simplify belchertown_root_url. Thx  Cheetah.
Cheetah was giving false positives on belchertown_root_url not being set.

My goal was to have people check this variable right away - as a way to catch them not reading the documentation - but Cheetah was finding the value not set in the first run which ends up raising the error - even though it'll get set on the last run of the generator.

Oh well (for now)...