Update so any Highcharts series data is accepted automatically
Without having to define every-single-highcharts-config, now we just accept it and add to the JSON array. JavaScript will handle that array automatically.

For example, you can have the following config:

```
        [[[chart2]]]
            title = Wind / Gust Speed
            [[[[windSpeed]]]]
            [[[[windGust]]]]
                zIndex = 1
                color = red
                name = Windy Gusty
            [[[[windDir]]]]
                type = scatter
                color = blue
```

This will give you wind speed, and wind gust as line charts, but on the same chart the wind direction as a scatter chart.

There are some downsides to this approach which revolve around the tooltips for the different data types (they won't work), and the yAxis' need to be defined for the 2nd data type (also won't work).