Introducing rainTotal, a special Belchertown skin observation vector
Special rainTotal Observation Type

The built in "rain" observation is really "bucket tips". The chart will show when the bucket tipped and that's it.

Using the special built in "rainTotal" observation - unique to Belchertown - will increment those bucket tips across the timespan, showing you the total rain for that timespan.