Dark mode overrides now persist throughout the session
A session is defined as the open tab or the browser.

When a visitor uses the toggle button to switch between light/dark themes, the auto mode is disabled. It cannot be re-enabled during the session.

If you refresh the page, the override will persist.

If you browse to other pages, the theme will persist if it's overridden.

When you close the tab or browser, auto theme will be re-enabled, and all persistent theme overrides will be removed until re-activated. Then this cycle repeats.

Updates #89