Better AJAX value rounding and clenaup (see description)
This update adds all the rounding values from weewx.conf to the weewx_data.json file for use with the MQTT AJAX update function. Previously a lot of these float values were hard coded, but now they'll follow what the user has set in their weewx.conf file.

This will also serve as groundwork for a future update on dynamic units.

For example before outTemp was hardcoded to 1 decimal value. But now if you want 3, just change it in weewx.conf and it'll show it on the page and on AJAX updates.

It also updates the station_observation values to use this new rounding array instead of the one I made specifically for that function previously.