Initial generator options - time delay your chart creation
By using generator = X, you can delay the chart generation for that time period. Useful for charts that don't need to be updated every archive interval (e.g. 5 minutes). So your year chart could be delayed to be updated once a day. This can help save your system from bogging down on chart load.

If the charts JSON file doesn't exist, it will be created on first run.

The timestamp of the file is used as the trigger on when to generate a new chart.

Options are:

hourly
daily
weekly
monthly
yearly
or any other time in seconds