Update belchertown.js.tmpl
edited the MQTT connection function to detect if the client is using https, and if so, connect to mqtt via tls.   (requires mqtt_websockets_port_ssl to be specified in config)

This change allows the skin to be served up without regard to http/https, works equally well either way.