belchertown.js.tmpl: Refactor precipitation handling
This makes an attempt to simplify the precipitation handling by removing
unnecessary distinction of cases. Snow and rain are both precipitation
and thus pop will be > 0 in both cases. There is also no real reason to
check for snowCM > 0 or snowIM > 0 in the querying code as the displaying code will decide
whether snow or rain is shown depending on the value of snow_depth
anyway.

This also implicitly fixes a bug where all forecasts following one with
snow also showed snow with the same amount as snow_depth was not cleared
until a forecast with pop == 0 was reached.